pyproject.toml
[tool.semantic_release]

version_variable = "werewolf/version.py:__version__"
upload_to_pypi = false
upload_to_release = false

[tool.poetry]
name = "werewolf"
version = "0.4.0"
description = ""
authors = ["Romain Pesche <romain@pesche.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
gunicorn = "^20.1.0"
django-guardian = "^2.4.0"
dj-database-url = "^0.5.0"
django-heroku = "^0.3.1"
bootstrap4 = "^0.1.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
